app/two_step_model/requirements-local.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.0.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
pathlib
argparse
ast
statistics
shutil
os
sys
traceback
